project-{topic}-pro/requirements.txt
# Core FastAPI and ASGI server
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1

# Database
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.1

# Pydantic for data validation and settings management
pydantic==2.7.1
pydantic-settings==2.2.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Caching and Rate Limiting
redis==5.0.4
fastapi-limiter==0.1.6

# Logging
loguru==0.7.2 # Enhanced logging capabilities

# Testing
pytest==8.2.0
pytest-cov==5.0.0
httpx==0.27.0
pytest-asyncio==0.23.6 # For testing async FastAPI apps

# Code Quality & Linting (Dev Dependencies)
black==24.4.2
isort==5.13.2
flake8==7.0.0

# Performance Testing (Dev Dependencies - for locust)
locust==2.25.1
```

```